backend/requirements.txt

aiohappyeyeballs==2.5.0
aiohttp==3.11.13
bokeh==3.4.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
attrs==25.1.0
babel==2.17.0
backports.tarfile==1.2.0
beautifulsoup4==4.13.3
bleach==6.2.0
blis==1.2.0
build==1.2.2.post1
CacheControl==0.14.2
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.6
comm==0.2.2
confection==0.1.5
crashtest==0.4.1
cymem==2.0.11
dataclasses-json==0.6.7
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.9
distro==1.9.0
docstring_parser==0.16
dulwich==0.22.7
spacy-model-en-core-web-sm==3.8.0
executing==2.2.0
fastapi==0.115.11
fastjsonschema==2.21.1
filelock==3.17.0
filetype==1.2.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
# google-ai-generativelanguage
google-api-core==2.24.2
google-auth==2.38.0
googleapis-common-protos==1.69.2
greenlet==3.1.1
groq==0.18.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
html2text==2024.2.26
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
idna==3.10
importlib_metadata==8.6.1
installer==0.7.0
ipykernel==6.29.5
ipython==8.32.0
isoduration==20.11.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keyring==25.6.0
langchain==0.3.20
langchain-community
langchain-core
langchain-google-genai
langchain-groq==0.2.4
langchain-ollama==0.2.3
langchain-text-splitters==0.3.6
langcodes==3.5.0
langgraph==0.3.25
langgraph-checkpoint==2.0.24
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.61
langsmith==0.3.11
language_data==1.3.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.1
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
murmurhash==1.0.12
mypy-extensions==1.0.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.3.2
notebook_shim==0.2.4
numpy==2.2.3
ollama==0.4.7
orjson==3.10.15
ormsgpack==1.9.1
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
phidata==2.7.10
pillow==11.1.0
pkginfo==1.12.0
platformdirs==4.3.6
playwright
poetry==2.0.1
poetry-core==2.0.1
preshed==3.0.9
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pyee==12.1.1
Pygments==2.19.1
pyppeteer
pyproject_hooks==1.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
pytz==2025.1
# pywin32==308
pywin32-ctypes==0.2.3
# pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.2.1
rank-bm25==0.2.2
RapidFuzz==3.12.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.29.0
semantic-split==0.1.0
Send2Trash==1.8.3
sentence-transformers==2.7.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.38
srsly==2.5.1
stack-data==0.6.3
starlette==0.46.0
sympy==1.13.1
tenacity
terminado==0.18.1
thinc==8.3.4
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.21.1
tomli==2.2.1
tomlkit==0.13.2
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
trio==0.29.0
trio-websocket==0.12.2
trove-classifiers==2025.1.15.22
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
uri-template==1.3.0
urllib3==1.26.20
uvicorn==0.34.0
virtualenv==20.29.2
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets
wrapt==1.17.2
wsproto==1.2.0
xxhash==3.5.0
yarl==1.18.3
zenrows==1.4.0
zipp==3.21.0
zstandard==0.23.0
python-multipart
google-genai
redis
httptools
pyppeteer
beautifulsoup4
selenium
webdriver_manager
langchain_community
html2text
playwright
semantic_split
zenrows
rank_bm25
fastapi
pyppeteer
google-generativeai
pyodbc


#file system
docling
cloudconvert
psycopg2
asyncpg
llama_parse
google-genai
PyPDF2
scikit-learn
python-multipart

# Visualization libraries
plotly
kaleido

email-validator
passlib[bcrypt]
python-jose
#pip install --upgrade google-generativeai
#pip install -U llama-index --upgrade --no-cache-dir --force-reinstall

# Voice Agent Dependencies
pipecat
loguru

# Pipecat framework with native audio streaming support
pipecat-ai==0.0.73

# FastAPI and WebSocket support
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==13.1
python-multipart==0.0.6

# Environment configuration
python-dotenv>=1.0.0

# Logging
loguru>=0.7.0

# Audio processing and VAD
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
soundfile>=0.12.0
librosa>=0.10.0
resampy>=0.4.0
onnxruntime>=1.16.0

# OpenAI Whisper for audio processing
openai-whisper>=20231117

# Native Audio LLM Services (for ultra-low latency)
openai>=1.0.0

# Google Gemini Live (via Pipecat)
google-generativeai>=0.3.0
grpcio>=1.50.0

# Fallback STT Services
deepgram-sdk>=3.0.0

# Fallback TTS Services  
cartesia>=1.0.0
elevenlabs>=1.0.0

# Protobuf for serialization
protobuf>=4.0.0

# Additional utilities
aiofiles>=23.2.0
requests>=2.31.0

# Audio processing dependencies
PyAudio>=0.2.11
pydub>=0.25.1
pyloudnorm>=0.1.0

# Machine learning and model dependencies
huggingface-hub>=0.16.0
transformers>=4.0.0
accelerate>=0.20.0
safetensors>=0.3.0


# Additional audio processing tools
edge-tts>=6.0.0
kaldiio>=2.17.0

# Optional performance optimizations
# For GPU acceleration (uncomment if you have CUDA)
# torch-audio-cuda
# torch-vision-cuda